Refactor: 상태 업데이트 로직 개선 및 스타일 경로 수정
